Created Line class and refactor Planet class to use Double members
Created a line class with the purpose of getting an easy way to test if
a point is part of a line given by two points.
Refactored the Planet and Galaxy classes to use type Double for
variables xPos, yPos and distance, this made it easier to implement the
Line class.
Added and changed some tests.